+++ _inc/client/recommendations/prompts/checkbox-answer/style.scss
