+++ _inc/client/recommendations/prompts/prompt-layout/style.scss
