+++ _inc/client/recommendations/prompts/site-type/style.scss
